Forget working
You can now forget the settings for a domain name by clearing the site name field in the popup.  I will remove the appropriate entry in database.domains and delete the bookmark with that domain name as the title.  In addition, it will remove the site name entry in database.sites if this site name no longer appears in the values of database.domains.

I also removed the instructions from ssp.html since they are now in their own file.