Fixed normalization problem
I wasn't properly updating site name changes when two domains mapped to the same site name and the capitalization of the two was different.  The change was a bit tricky since I still want to keep the user's capitalization.

I also introduced a boolean for test mode.  In bg.js it turns on logging and uses SitePasswordDataTest.  (I was always forgetting to set the latter.)  In ssp.js, it turns on logging and turns off auto close of the popup window.