Don't find userid field if there's more than one password field
The heuristic for finding the userid field is that it is the first visible input field above the password field, which works in almost all cases.

One case that is annoying is change password forms that have multiple password fields.  Some pages allow showing the password, which makes the field look like an input field of type "text."  The result is that the heuristic replaces the password with the userid.  Since these pages don't typically have a userid field, the fix is not looking for one if there's more than one password field.