Fix problem where password gets replaced by userid
Some sites let you see your password, which changes the input type from password to text.  The result is that the heuristic for finding a userid field actually finds a password field with a visible password and replaces the password with the userid.

To solve this problem, I remember the maximum number of password fields found and test that to see if I should look for a userid field.  This approach assumes that the password fields are not set to show the password by default.  If that assumption is wrong, I won't find the password fields.