Don't do anything on SitePassword web page
You wouldn't normally open the SitePassword web page if you have the extension running.  If you do, you'll see "Click SitePassword" in the master password field, which is misleading.  I put a marker div in the SitePassword HTML so the content script knows not to do anything in that case.