Add strength meter
This commit adds the same strength meter as in SitePasswordWeb with a couple of display differences.

The strength meter is put in the master password field to save some real estate.  Because the user's master password may overlap the meter, the meter is hidden when the master password is either empty or visible.

There is a potential, but highly unlikely, chance that generated password will no longer be generated, since this change rejects passwords that are too "weak." where weak depends on the size of the master password.

This commit also fixes a bug introduced when I changed from a mix of "pwlength" and "length" to always use the former.  I had missed two places.