Made color for weak more readable
The color chosen for weak passwords is barely readable on the web page, but not readable on the popup because it is so much smaller.

I made that color a bit more saturated, but then it was too close to the color for bad passwords, so I changed that color a bit, too.