Fix a rare bug
Rarely on starting, bg.settings doesn't have sitename or username set to the default of "".  The result is an error dereferencing them.  I have been unable to capture the case when it occurs, so I haven't been able to debug the problem.

I think the problem might be in bg.js, so I double check.  Just in case the bug isn't there, I check again in generate.js because that's where the error gets reported.