Add alert for HTTP login pages
HTTPS is nearly universal and may be universal for login pages.  In case it is not, the user should be notified of the danger before filling in the password field on an HTTP page.  Unlike most such dialog boxes, this one will appear rarely enough that users are likely to pay attention to it.