Two fixes for HTTP warning
Dale has an HTTP login page for his blog and doesn't want to see the warning popup every time he visits the page, so I only show it to him when the userid field is empty.  That way he'll only get the warning until he's set up the site.

I was only injecting the content script for HTTPS pages.