Alert user when password is put on the clipboard
It's not a good idea to leave a site password on the clipboard for any longer than necessary.  Some password managers clear the clipboard after a fixed interval.  Picking that interval is a problem.  Too short, and the user might not be done pasting.  Too long, and the user might have put something else on the clipboard.

I've chosen a different strategy.  When the user pastes a site password, I add a badge to the SitePassword icon and change the title to say a password might be on the clipboard.  I'd prefer for the user to be able to clear that value by double clicking on the SitePassword icon, but I can't figure out how to make that work.  Instead, I've added a "Clear Clipboard" icon.  Clicking it writes "Password cleared" to the clipboard and resets the badge and title to their original values.