Don't do any work if the tab isn't active
There are two places where I do work that doesn't need doing if the tab isn't active.  The first is a setInterval looking for updates to the CSS.  The second is in the mutation observer.

While it's theoretically possible that a page will make a password field visible when its tab is inactive, that's highly unlikely.  This change makes SitePassword more polite by not doing any work in those two cases.