Added instructions to popup
Going to another page for the instructions is annoying for the user and a bit of a burden for me.  Adding the to the popup is simpler for both.  I did that by copying the text from instructions.html, which I deleted, and adding the script required for the accordion.

There are also some updates to the instructions.