Improve phishing warning message
I was asking the user to look at the address bar when checking for phishing.  This change puts the domain name in the warning.

There are two advantages.  First, I control the font it appears in, which makes certain kinds of spoofing harder.  Second, I'll be able to show the domain name of an iframe once I get that working.