Fix phishing vulnerability
You get a phishing warning when providing a site name that was used for a different domain name.  That warning said to check the address bar to make sure you weren't being phished.  That won't help if the password field is in an iframe.

The new warning message puts the offending domain name in the message.  It will be what's in the address bar if there's no iframe and the iframe's domain name if there is one.

There is still a potential problem.  I keep a map of the address bar to iframe domain names in the service worker.  That map is lost if the service worker goes idle between when it is set up and when it is used.  I can write it to local storage, or I can find a way that doesn't need state that survives the service worker shut down.