Fixed problem with service worker losing pwdomain info
My original fix for the bad phishing message used a hashmap in the service worker.  Unfortunately, the service worker loses its memory when it is suspended.  That probably wouldn't happen in the short time between when I set up the hashmap and when it gets used, but it could happen.

This commit puts the domain name of the password field in bg.settings, which gets saved and restored by the service worker.  Note that I save pwdomain even if the password field is not in an iframe.  Hence the check to see if the two are the same.