Deal with pwdomainname != domainname
The pwdomainname is something the user should only see in a phishing warning, so I don't want it to show up in the bookmarks.  However, I need it in the database.  That's what these changes do.

However, I currently don't have a way to delete the pwdomainname entry when the user clears the site name.  I can't just remove it because there might be associated with another domain name.

I'm leaving it in test mode until I figure out what to do.