Turning off test mode
because everything seems to be working.  The last concern was dealing with iframe domain names when forgetting a domain name.  I unconditionally delete the iframe entry when I delete the page domain name entry even if there are other page domain names that use the same iframe domain name.

It turns out that isn't a problem because the next time I load the page with the domain that I didn't delete, the iframe entry gets recreated.