Fixed a long standing problem
Open the popup on a page with a password field.  Mouse out of the popup, and open it again.  It says it couldn't find a password field.  The problem is that I was overwriting what the service worker knew about the number of password fields.

I also updated the storyboard yet again.  Please, someone stop me before it's too late.