Fix auto close of popup
When you mouse out of the popup, it closes even if you mouse back in.  The commit cancels the close if you do that.

I've set the close timeout to be 3/4 second.  If you click on the password field faster than that, you have to click again to get your password.  It's a bit annoying but less so that having the popup close when you accidentally mouse out and then back in again right away.