Fixed a dumb mistake
I was assigning the DOM element containing the site name instead of the value of that element to the sitename variable.  I didn't find the error until I turned on logging for another reason because sitename got reassigned to the value before I ever used it in anything the user would see.