Changed Is Password Field Visible
I did a check to see if window.computedStyle() always gets the visibility check right.  Surprisingly, it does not.  I verified that when it's wrong on one site I also have to check the offsetParent.  I did find the password field on all my test sites without the ariaHidden check.

Also, more updates to the slides.