Include opacity in visibility test
There is no good reason to set the opacity on a password field.  You can safely assume that something clickjacky is going on if it's set.  Just in case someone does set the opacity for a legitimate reason, I'll call it visible if the opacity is high enough to be noticeable.