Create separate bookmark for just common settings
DOES NOT WORK: I need to build the database from these new bookmarks.

The bookmarks were holding the settings twice, once in those with numbers for titles and once in those with domain names for titles.  The result was a complicated process of stitching together the database from multiple numbered bookmarks since Google limits the size of what gets synched.

This commit removes the redundancy when creating the bookmarks.  The CommonSettings bookmark holds those items that are independent of the domain name, and the other settings are held in the bookmark with the title corresponding to the domain name.