Set updateTime if changed
I need to know when the settings for a domain name change if I'm going to automatically merge them.  I could set it when I get settings from the popup, but I've chosen to do it when persisting the database.  As an added bonus, I can skip updating the bookmark if the settings haven't changed.