Fixed problem when 2 domains share settings
I thought there was a bug, because the first to be created showed the domain name of the second to be created.  Of course, that's expected since the domain name is in the settings, and the settings are shared, so only one can be there.

The bug is that I was using the value in the settings for the domain name to be put into the database instead of the bookmark title.  That prevented the second bookmark from being created.

I also had an if condition wrong, so I wasn't setting pwdomain when I should have.