Remove legacy bookmarks
I used to store the full database in a set of bookmarks with numbers for titles.  Since the database no longer does that, those bookmarks are never looked at.  That means the easiest way to update an existing database is to simply delete those bookmarks the first time they are encountered.  The update is completely transparent to the user unless they are looking at the bookmarks folder.