Ignore blank site name
In the old code you would forget the settings for a domain by clearing the site name field.  That wasn't obvious and was hard to explain.  With the new approach, you just delete the corresponding bookmark.

I also changed the test to see if I need to update the bookmark to use sameSettings(), which I updated to ignore the domain name field.  That field can change if there are multiple domains that share settings.