Fixed problem with pw fields in iframes
When a password field is in an iframe, the first password calculation was using the default settings.  Subsequent calculations were correct.

Because synchronization can make a new bookmark appear on an existing one vanish at any time, I have to rebuild the database every time the service worker receives a message.  One of those messages is from the content script that tells the service worker the domain of the iframe.  Another is from the popup with the settings from the filled in form.  The latter doesn't include the domain of the iframe.  The result is several places where I was replacing the contents of the first message with the default settings.  However, there are times when the default settings are needed.

The fix is to check to see if they're needed before using the default settings.