The case of the reappearing bookmark solved
After deleting a bookmark and reloading the corresponding page, the bookmark reappeared, and the popup settings were filled in.  Of course, that shouldn't happen.

The problem is that I cleared the database before rebuilding it, but I hadn't cleared bg, which meant they were put back into the database before it was saved.  Hence, the reappearing bookmark.

When the problem first appeared, the bookmark was created twice.  My code correctly deleted the duplicate, since the two had identical URLs.  The problem came from findpw.js, which sent page info on document.onload() and window.onload().  I fixed that by sending only on the former.