Got rid of update time
I add updateTime to bg.settings when I was planning to automatically merge duplicate bookmarks.  The plan was to only leave the most recent, but that's not the only possible policy.

I've decided to leave picking which of the duplicates to keep up to the user.  I hope the alert message provides enough information.