Fixed a long standing but rare problem
Load a page with a password field, then open the popup.  If you close the popup and reopen it, you'll rarely get the "no password fields" warning even though you didn't get it the first time.  I thought it was an async problem that I didn't know how to reproduce, so I put off fixing it.

I now know that the problem is the service worker stopping between the two openings of the popup.  It forgets the pwcount and the associated URL.

My fix is to put that data in session storage when the service worker gets the onload message and retrieve it when the popup asks for the metadata.

There is a 5 MB limit, but I only store the data for pages that have a password field, and it gets cleared when the session ends.