Fixed broken phishing warning
In the process of finishing getting the figures right for the talk, I saw that I wasn't getting the phishing warning from allantheguru because getMetatat() was sending an empty database to the popup.

The problem is that there's an async call to get the saved password count.  That means getMetadata() returns before the async callback runs, but the database gets set to its default value in the interim.

The fix is to assign the database in a local variable, which is the state that gets captured by the closure.