Neglected to send response to getMetadata request if no saved data
The popup stopped showing the "no password fields" found message.

When the popup opens, it asks the service worker for metadata, which includes data saved when the content script loads.  However, that data never gets saved in two cases.  If there are no visible password fields on the page, the content script never sends the "count" message.  The other case is for file:/// URLs, which don't get a content script.

I was assuming that the absence of savedData meant the service worker didn't need to respond to the getMetadata message.  That's clearly false, because the popup will still ask for the metadata.

This fix sends default metadata in the two cases where the message was not previously being sent.