Recover jQuery bullet point
I gave the talk last night and realized that I hadn't mentioned jQuery.  The reference was on two slides.  One I had moved to backup, and the other one had the comment in the speaker notes but not in the bullet points.  It's now on the slide that includes the code size bullet.