Actually save common settings
While I was setting the common settings, clear master password on use and hide site password by default, in the popup, I wasn't actually saving them in the database due to an await error.

The problem is that the "database" variable is global, so its current state doesn't get captured in the implicit closure following the await.  The fix is to assign database to a local variable just before the await and use the local variable after it.