Clear browser cache when the extension updates
Pages need to be reloaded following an update to the extension, but the page may come from the cache.  The result may be confusing messages, such as "Can't find password field" when it's plainly visible.

Fortunately, there are APIs for detecting an update and for clearing the browser cache.  I check for update every time the service worker starts and clear the cache if there was an update.