Add link to web page in popup
If you need a password when you're away from your machine, you can get it from https:// sitepassword.info.  But how do you remember that URL?  This commit changes the title on the popup to be a link to that web site.

There's no guarantee users will remember the URL, but there's a better chance if "SitePassword.info" is visible every time they open they open the popup.

I also added a statement to the downloaded table of settings.

This problem screams for a user study.  I just wish I could do one.