Handle file:/// URLs
I was returning a blank domain name for file URLs.  I now return the full path to the file as the domain name.  I think that makes sense, since that's the thing that uniquely defines the file in case it contains a password field that you want to use.  I also decided to deal with mailto URLs.

All other commonly used protocols separate the protocol from the rest with ://.