Remove bankofamerica.com code
bankofamerica.com clears the password and username fields a second or two after the page loads, so I put in code to insert the values again.

That causes a problem when you need to paste in a password, such as the way some sites do a "Forgot password" as does https://reg.rainfocus.com/flow/rsac/365/selectCFS/login.  In that case the computed password replaces the pasted one, which is not what you want.

I believe it's harder for the user to figure out what to do in the latter case, so I removed the bofa code.