Fixed disable of providesitepw check box
You should only be able to provide your own site password if there is a calculated site password to use as the encryption key.  That condition requires at least a master password, but I also insist on a site name and username.  I had forgotten to test for the master password.