Fixed module problem
Firefox kept complaining that I could only load modules from the top level, but I thought that bg.js was at the top level.  However, Firefox wrapped it in a default .html file.  The fix was to create my own .html file to load it.  This was something I had to do on Chrome Manifest 2, but that was so long ago that I had forgotten about it.