Unified code for Chrome and Firefox
There are only a few differences in the code for Chrome and Firefox, so it makes sense to have the same code with conditionals for which version to use.  This version works in both browsers, but Firefox gives a console error about needing a content security policy for sitepassword.info when I open the popup.

You have change the name of the manifest file from manifestChrome.js or manifestFirefox.js to manifest.js.