Changed ids for username element
Doug Crockford is having a problem with browser autofill interfering with SitePassword.  He suggested that the browser might be inferring what to do with an element based on its id or label.  I changed the id to see if that  helps, but I don't know what to do about the label and still make it meaningful.

This commit also contains a set of 8 candidates for a new logo.