Basic functions work for Safari
The Chrome bookmarks API gives you an array of bookmark objects.  Safari doesn't support that API, so I have to use browser.sync.storage, which gives you an object, each element corresponding to a bookmark.  The trick that made this work was using Object.values() on the object and making sure each entry has a "title" field.