Make bookmarks available to copy
There may be times when you're at a machine that doesn't have your bookmarks.  In that case you have to use the web page to get your passwords by manually entering your settings.

With this change, you can copy and paste the appropriate bookmark if the file you downloaded with your settings is available, e.g., Dropbox or iCloud.