Open a page instead of downloading settings
Safari doesn't support the download API, so I decided to open a new tab with the table.  Unfortunately, Chrome doesn't allow you to do that at the top level.  I have to open it in an iframe.  The problem is that I have no control over the protocol and domain of the links in <a href= when I do that, which led to changes in SitePasswordWeb.