Added a section in README for developers
The "name" in the manifest shows up in the extensions manager page and as a tooltip on the icon.  I was using that field to distinguish the different versions of the manifest file, but it was annoying.  As far as the user is concerned it us SitePassword, not SitePassword for Chrome.

I don't know where the "short name" is used, but putting which browser it's for there hasn't turned up anything I don't like.