In case Safari implements bookmarks API
Safari does not currently support the bookmarks API, but it might in the future.  If I don't do anything, Safari users will lose their settings.  Even worse, they'll get confusing results if they upgrade their browser on one machine but not another.  My solution is to always set sync storage but use bookmarks if they are supported.

I have to be compatible with both browsers that currently support the bookmarks API and those that don't.  First, I get the sync data if it exists.  Next, I get the bookmarks if they exists.  If neither exists, then this is the first run.  If both exist, then I already did the necessary setup.  If one exists but the other doesn't, and the browser supports the bookmarks API, then I build the missing one from the one that exists.