Fixed download problem
Chrome no longer allows a top level domain to be a data type, such as data:text/html, and Safari doesn't support downloads from an extension.  I can open a tab with the formatted table of settings, but it's domain is about:blank, which doesn't allow SaveAs.

The answer is to create an "a" element with the href being the data representing the table, giving it a "download" attribute. and then automatically clicking on it.