Get rid of annoying error
If a component sends a message, and the receiving side doesn't send a response, you get a

Uncaught (in promise) Error: A listener indicated an asynchronous response by returning true, but the message channel closed before a response was received

error.  Nothing bad happens, but it might mask a more serious error.