Two changes to datalist
Hide the down arrow to the right of an input field that has a datalist.  It's not needed, and it overlapped the clipboard icon in the user name field.

Some browsers use the value in a datalist option; some use the innerHTML, so I decided it was best to provide both.