Fixed problem with forgetting domain
On April 10 in commit 346d4d726e0e9a1db61cb1ffd9f8f0025161b670 I made some changes to deal with the case of Safari implementing the bookmarks API at some point in the future.  (Read the commit notes for further explanation.)  In the process, I broke the ability to forget a domain.

The problem is that I was using the Safari approach of sync storage for the settings and setting the bookmarks from that.  That meant a deleted bookmark would be reconstituted from sync storage.  Rather than try to come up with a better solution for Safari now, which might take some time, I reverted to using sync storage only for Safari.