Update no passwords field message
It's not uncommon for SitePassword to say it couldn't find any password fields even though it can.  The problem is that the page needs to be reloaded after a SitePassword update.  I could try reloading the page when no password field is found, but that could interfere with the user's activities.  Instead, I'm changing the message to say to try reloading.