Work around known Firefox bug
Firefox triggers a mouseleave event when you move your mouse into the datalist.  That causes the popup to close.  I fixed that problem by checking which element the mouse is in when a mouseleave is triggered.  If there's no element, then the mouse is outside the window, and the popup can close.  If not, the mouse is in the datalist, and the popup should not close.