Indicate on popup if password was put on clipboard
I put a warning on the SitePassword icon that you might have left a site password on the clipboard.  Doug Crockford didn't know what that meant and suggested I put something on the popup.

When you click the copy icon on the site password field, I change the badge on the icon as before and change the logo in the popup window to match.  Both are cleared when you click Clear Password.