Fix badge on icon on toolbar
When you copy a site password to the clipboard, I change the icon on the extensions toolbar using setBadge() as a reminder to remove it when you're done.  Unfortunately, Chrome and Firefox set the badge differently.  Chrome puts in on the bottom half; Firefox, on the top.

That wasn't a problem until I changed the icon on the popup to match, because they match on Chrome but not on Firefox.  My fix is to create a second icon and put the appropriate one on both the toolbar and the popup.