Get rid of document.write()
I need to create either a new window, tab, or downloaded file when the user clicks Download Site Data.  If I create a new window on Chrome, I can't do a SaveAs because the top level is a data URL.  Opening a tab does work.  However, Safari doesn't support the tabs API.

Even though my use of document.write() is safe, someone looking at the source code wouldn't immediately know that.  In order to get rid of it I removed the code that built the HTML by concatenating strings and replaced it with code that builds the page by creating elements and appending them.  This approach is better anyway because open and close tags always match.

Still needs to be tested in Firefox and Safari.