Fixed handling of special characters
I wasn't escaping special characters when going back and forth between settings and bookmarks.  As a result some characters were converted to codes, e.g., "<" became %3E in the special character field of the popup.

Since I already had code to convert between strings and arrays of character codes from implementing user provided passwords, I decided to use that instead of escaping.  The special characters are stored in the bookmark as an array of character codes.  The special characters in the settings are represented as strings.