Fix service worker keep-alive for Firefox
Firefox does not support storage.session, so I have to use sessionStorage.  The former survives a restart of the service worker; the latter does not.  As a result, SitePassword on Firefox forgets your master password every 20 seconds or so.

Chrome has a strict 5 minute timeout for a service worker even if it's actually receiving messages.  It appears that Firefox does not have such a limit.  Hence, I can keep the service worker alive by having the content script send it a message to it every 10 seconds.

I only need one tab to send the message, but every tab will.  I could use document.hidden to only send from active tabs, but then I'd lose the master password if no tab running the content script was active.  The overhead is low enough that it's not a performance problem to send from all tabs.