Undo conditional keep alive
For some reason only sending the keep alive messages on Firefox and not Chrome didn't keep the service worker alive.  I verified that the service worker got the message every 10 seconds, but the super password kept disappearing.  Taking out the conditional part keeps the service worker running, but it means the messages will also be sent on Chrome.  The overhead is so low it doesn't matter.