Turn off keep alive if using storage.session
The data in storage.session survives service worker shutdown, but the data in sessionStorage does not.  In the latter case, I keep the service worker alive by sending a keep alive message from the content script to the service worker every 10 seconds.

If I continue to send the keep alive messages on Chrome after an update, I get "Extension context invalidated" errors, so I stop sending these messages if storage.session is available.  I don't have that problem on Firefox, which appears to re-initialize the content script.  (The debugger remains connected to the un-re-initialized instance.)