Change popup UI and some minor clean up
1. Simplify handling of new defaults by getting rid of the global variable newDefaultSettings.

2. Fixed a bug in suppressing keep alive messages.  I had inadvertently shadowed the keepAlive interval variable.

3. Remove a duplicate of two object values.

4. Popup UI changes
a. Changed the UI to make clearsuperpw on use had hidesitepw to be global settings.
b. Moved Provide your own site password from just under the site password field to the top of the settings area because I'm trying to discourage its use.
c. Replaced the box around the settings that can be saved with <hr>