Fixed a phishing related bug
The user gets a phishing warning when entering a previously used site name for a different domain.  If the user says this domain is for the same account, then the form gets filled in and the site data is sent from the popup to the service worker, which creates a bookmark for the new domain.  I don't want to send any data to the service worker if the user mouses out before clicking the same account button.

The problem is I was running the isphishing() check in the popup before the new bookmark could be created.  As a result, the service worker never got the data, and the bookmark was never created.

Fortunately, I have another test that tells me whether there's a phishing warning showing in the popup, namely whether the user name field is empty or disabled.  This test also fixes a minor usability issue that arises when the user mouses out before filling out both the site  field.