Keep setting open when choosing same account
One option you have when you get a phishing warning is to say that this domain is for the same account as for another domain.  When click that button, the settings get filled in from the other domain's bookmark.  The settings then close, generating a mouseleave event.  That's annoying because you may want make some other changes.  This commit keeps the settings open until you mouse out of the popup.