Fix autoclose test
When you mouse out of the popup, it sends the settings and super password to the service worker, something you don't want to do if you might be at a phishing site.

My previous test worked, but it didn't remember your master password if you entered it but not your site and user names.  This test does what I want by seeing if the phishing warning is visible.  Note the need to make sure the site name is not focused, which is needed to make the phishing warning visible.