Start on new way of dealing with options
Different fields have different things you can do.  For example, you can copy your user name, but you can copy and hide your site password.  Previously, that was done with an icon in the field for each thing you could do.  That's limiting and makes the UI look busy.

This branch tries out a different approach suggested by Doug Crockford in which each field has a 3-dot icon that expands when you hover over it to show a menu of options you have for that field.

THIS VERSION IS NOT FUNCTIONAL.