Don't create character set in each iteration
The chars string is used to convert the hash into a site password.  For example, it consists only of alphanumerics if special characters aren't allowed.  That set depends only on the settings, which are independent of the number of iterations.  In spite of that, I was constructing the list on each iteration.