Next try at help
I didn't like using tooltips for the help because you have to click on the other menu options.  This version pops up a box with help for the specific field when you click on Help in the 3 dots menu.

I'm not happy with the way it looks or the text of the help, but the approach is better than using tooltips.