Fixed error in merge
I merged the changes made in main that put the instructions next to the popup form rather than below it into this branch, but I mistakenly accepted a partial update.

The full change doesn't let you focus on the domain name field, which makes sense for the code on the main branch but not for this one that shows 3 dots when a field is focused.