Forget domainname working
When you click Forget in the domain name field help menu, the popup deletes the corresponding bookmark and asks the service worker for the new settings for the domain name, now the default settings, and the new database, which no longer has an entry for the domain name.  As a nice side effect, the site name won't be included if there are no domain names that use it.