Clear login form when forgetting domain name
I was correctly clearing the form in the popup when you forget the settings for a domain name, but the login form was left with the values there before the forget, which looked odd.

I also needed a change in the content script.  Previously, I couldn't replace what was in a field with a blank.  Now I can.