Fixed a newly introduced bug
There are times when I want to clear a field if it has something in it and times when I don't.  I've been sending an empty string in both cases, but that means I can never replace the contents with an empty string.  I introduced a bug where the password field disappeared when you clicked on it in my attempt fix the first problem.

The actual solution was much simpler.  Send "" to leave what's there; send " " to replace it with the empty string, which works because I always trim the input.