Show settings on the side
A Chrome extension popup can only be a max of 600px high.  That means the form has to be pretty compressed to show the settings within that amount of space.  However, the extension can be as much as 800px wide.  That means putting the settings next to the base form leaves plenty of vertical room  I haven't added that space yet, though.