Show 3dots on mouseover and move menu
I was showing the 3dots menu when a field gets focus.  I now show it during mouse over.  The 3 dots showing as you move across the form isn't annoying.  It is annoying if you move your mouse over the 3 dots icons, but that's easy to avoid.

I've also positioned the menu item that is most likely to be the one you want directly under the mouse.