Fix yet another async error
The error, which showed up frequently do to an async error, is still occurring, but less often.  That's because there is nothing to put into bg the first time the service worker starts.

The error arises when I invoke persistMetadata() before setting up bg, which was initialized to {}.  The fix is to initialize bg to {superpw: "", settings: defaultSettings}.