Fix problem with instruction icons
The instructions are an accordion, which means one section closes when another one opens.  Dale notice that the open/close icons weren't being reset properly when I closed a section.  This commit fixes that problem.

In testing, I also found that a couple of sections were not closing, which I fixed, too.