Preconfirmation for Forget
Forget is dangerous, since it can't be undone.  Forget of a user name is particularly dangerous, because it may forget a LOT of settings.

Confirmation dialog boxes don't help much because people click through them.  Preconfirmation is better.  It's like the cover on the "launch nukes" button.  In this case, you click the Forget icon, which brings up a Forget box showing you the domain names to be forgotten.  At this point you can Cancel.