Test new approach to help
I currently have specific help text for each field in the main menu.  A different approach would be to use the appropriate section in the instructions.

The argument for is consistency; you see the exact same explanation if you click the help icon or the instructions icon.

The argument against is focus; the individual help items can be less comprehensive than the instructions