Try help text with a More button
There's a tension between providing a short bit of help and providing more comprehensive information.  Adding a More button to the help text addressed this issue.

This code is a trial and only works for the super password field.  It also doesn't work if you click More an even number of times.