Change color of pressed button
Red is far too jarring.

I also changed the look of the button when you hover over it.  Now the border is enhanced rather than vanishing.  I had to change the padding, too, or else the page appeared to move slightly when you moused over the button.  The movement was slight but annoying.